added user test
