+++ serverless/packages/core/src/parseJwt.ts
