+++ serverless/packages/functions/src/user/get_user.ts
@@ -1,13 +1,13 @@
 import { APIGatewayProxyHandlerV2, APIGatewayProxyEventV2 } from "aws-lambda";
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
 import { DynamoDBDocumentClient, GetCommand } from '@aws-sdk/lib-dynamodb';
-import parseJwt from "@dlp-sst-app/core/parseJwt";
+import parseJwt from "../../../core/src/parseJwt";
 
-export const handler: APIGatewayProxyHandlerV2 = async (event : APIGatewayProxyEventV2) => {
+export async function handler<APIGatewayProxyHandlerV2>(event : APIGatewayProxyEventV2) {
     if (event)
     {
         const user_id: string = parseJwt(event.headers.authorization ?? "")["user_id"];
-
+        console.log(event.headers.authorization ?? "")
         const client: DynamoDBClient = new DynamoDBClient({});
         const docClient = DynamoDBDocumentClient.from(client);
