+++ serverless/packages/functions/tsconfig.json
@@ -5,8 +5,8 @@
     "moduleResolution": "node",
     "baseUrl": ".",
     "paths": {
-      "@dlp-sst-app/core/*": ["../core/src/*"],
-      "@dlp-sst-app/functions/*": ["../functions/src/*"]
+      "@dlp-sst-app/core/*": ["../core/*"],
+      "@dlp-sst-app/functions/src/*": ["./src/*"]
     }
   }
 }
